refactor: 更新RTT平台编译文件
