+++ SConscript
@@ -6,12 +6,12 @@ cwd = GetCurrentDir()
 
 # The set of source files associated with this SConscript file.
 src = Glob('common/*.c')
-src += Glob('pahoMqtt/*.c')
+src += Glob('coreMqtt/*.c')
 src += Glob('mqttclient/*.c')
 src += Glob('platform/rtthread/*.c')
 
 path = [cwd + '/common']
-path += [cwd + '/pahoMqtt']
+path += [cwd + '/coreMqtt']
 path += [cwd + '/mqttclient/include']
 path += [cwd + '/platform/rtthread']
